requirements.txt
asgiref==3.7.2
chardet==5.2.0
colorama==0.4.6
Django==5.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
fire==0.5.0
fonttools==4.47.0
lxml==4.9.4
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
pdf2docx==0.5.6
pdfkit==1.0.0
Pillow==10.1.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pywin32==306
reportlab==4.0.8
six==1.16.0
sqlparse==0.4.4
termcolor==2.4.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
XlsxWriter==3.1.9
